Export enhanced
